pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "text2video"
version = "0.1.0"
description = "Seedance-like AI video generation scaffold with AWS-lite infrastructure"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.39.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "pillow>=11.0.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "python-dotenv>=1.1.0",
  "uvicorn>=0.35.0",
]

[tool.setuptools]
package-dir = {"" = "packages/python"}

[tool.setuptools.packages.find]
where = ["packages/python"]
